Refactor end-to-end tests to a separate module.
Also implement testing for progressive decoding.
