+++ jxl_macros/src/lib.rs
@@ -718,57 +718,3 @@ fn derive_jxl_headers_inner(input: &DeriveInput) -> syn::Result<TokenStream2> {
 pub fn noop(_attr: TokenStream, item: TokenStream) -> TokenStream {
     item
 }
-
-#[cfg(feature = "test")]
-#[proc_macro]
-pub fn for_each_test_file(input: TokenStream) -> TokenStream {
-    use std::{fs, path::Path};
-    use syn::Ident;
-
-    let fn_name = parse_macro_input!(input as Ident);
-    let workspace_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("..")
-        .canonicalize()
-        .unwrap();
-    let root_test_dir = workspace_dir.join("jxl").join("resources").join("test");
-    let conformance_test_dir = root_test_dir.join("conformance_test_images");
-
-    let mut tests = vec![];
-
-    for test_dir in [root_test_dir, conformance_test_dir] {
-        for entry in fs::read_dir(&test_dir).unwrap() {
-            let entry = entry.unwrap();
-            let path = entry.path();
-            if path.extension().is_some_and(|ext| ext == "jxl") {
-                let relative_path = path
-                    .strip_prefix(&test_dir)
-                    .unwrap()
-                    .to_string_lossy()
-                    .replace('/', "_slash_");
-                let test_name = format!(
-                    "{}_{}",
-                    fn_name,
-                    relative_path.strip_suffix(".jxl").unwrap()
-                )
-                .replace(|c: char| !c.is_alphanumeric() && c != '_', "_");
-                let test_name = Ident::new(&test_name, fn_name.span());
-                // Path relative to workspace root.
-                let rel_path = Path::new("..")
-                    .join(path.strip_prefix(&workspace_dir).unwrap())
-                    .to_string_lossy()
-                    .into_owned();
-                tests.push(quote! {
-                    #[test]
-                    fn #test_name() {
-                        #fn_name(&Path::new(#rel_path)).unwrap()
-                    }
-                });
-            }
-        }
-    }
-
-    quote! {
-        #(#tests)*
-    }
-    .into()
-}
